moving more code to ANTLR module
