+++ core/src/main/kotlin/com/strumenta/kolasu/parsing/Parsing.kt
@@ -7,7 +7,6 @@ import com.strumenta.kolasu.validation.Issue
 import com.strumenta.kolasu.validation.IssueSeverity
 import com.strumenta.kolasu.validation.IssueType
 import com.strumenta.kolasu.validation.Result
-import org.antlr.v4.runtime.*
 import java.io.BufferedInputStream
 import java.io.ByteArrayInputStream
 import java.io.File
@@ -63,13 +62,6 @@ open class KolasuToken(
 ) : Serializable
 
 /**
- * A [KolasuToken] generated from a [Token]. The [token] contains additional information that is specific to ANTLR,
- * such as type and channel.
- */
-data class KolasuANTLRToken(override val category: TokenCategory, val token: Token) :
-    KolasuToken(category, token.position, token.text)
-
-/**
  * The result of lexing (tokenizing) a stream.
  */
 class LexingResult<T : KolasuToken>(
@@ -96,39 +88,11 @@ class LexingResult<T : KolasuToken>(
     }
 }
 
-class FirstStageParsingResult<C : ParserRuleContext>(
-    issues: List<Issue>,
-    val root: C?,
-    code: String? = null,
-    val incompleteNode: Node? = null,
-    val time: Long? = null,
-    val lexingTime: Long? = null,
-) : CodeProcessingResult<C>(issues, root, code) {
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (other !is FirstStageParsingResult<*>) return false
-        if (!super.equals(other)) return false
-
-        if (root != other.root) return false
-        if (incompleteNode != other.incompleteNode) return false
-
-        return true
-    }
-
-    override fun hashCode(): Int {
-        var result = super.hashCode()
-        result = 31 * result + (root?.hashCode() ?: 0)
-        result = 31 * result + (incompleteNode?.hashCode() ?: 0)
-        return result
-    }
-}
-
-class ParsingResult<RootNode : Node>(
+open class ParsingResult<RootNode : Node>(
     issues: List<Issue>,
     val root: RootNode?,
     code: String? = null,
     val incompleteNode: Node? = null,
-    val firstStage: FirstStageParsingResult<*>? = null,
     val time: Long? = null
 ) : CodeProcessingResult<RootNode>(issues, root, code) {
 
@@ -193,46 +157,3 @@ interface KolasuLexer<T : KolasuToken> : Serializable {
     fun lex(file: File): LexingResult<T> = BufferedInputStream(FileInputStream(file)).use { lex(it) }
 }
 
-fun Lexer.injectErrorCollectorInLexer(issues: MutableList<Issue>) {
-    this.removeErrorListeners()
-    this.addErrorListener(object : BaseErrorListener() {
-        override fun syntaxError(
-            p0: Recognizer<*, *>?,
-            p1: Any?,
-            line: Int,
-            charPositionInLine: Int,
-            errorMessage: String?,
-            p5: RecognitionException?
-        ) {
-            issues.add(
-                Issue(
-                    IssueType.LEXICAL,
-                    errorMessage ?: "unspecified",
-                    position = Point(line, charPositionInLine).asPosition
-                )
-            )
-        }
-    })
-}
-
-fun Parser.injectErrorCollectorInParser(issues: MutableList<Issue>) {
-    this.removeErrorListeners()
-    this.addErrorListener(object : BaseErrorListener() {
-        override fun syntaxError(
-            p0: Recognizer<*, *>?,
-            p1: Any?,
-            line: Int,
-            charPositionInLine: Int,
-            errorMessage: String?,
-            p5: RecognitionException?
-        ) {
-            issues.add(
-                Issue(
-                    IssueType.SYNTACTIC,
-                    errorMessage ?: "unspecified",
-                    position = Point(line, charPositionInLine).asPosition
-                )
-            )
-        }
-    })
-}
